+++ composer/src/main/java/com/holonplatform/datastore/jdbc/composer/dialect/OracleDialect.java
@@ -134,7 +134,7 @@ public class OracleDialect implements com.holonplatform.datastore.jdbc.composer.
 			return Optional.of("FOR UPDATE NOWAIT");
 		}
 		if (timeout > 0) {
-			return Optional.of("FOR UPDATE WAIT " + ((timeout < 1000) ? 1 : Math.round(timeout / 1000)));
+			return Optional.of("FOR UPDATE WAIT " + ((timeout < 1000) ? 1 : Math.round((float)timeout / 1000)));
 		}
 		return Optional.of("FOR UPDATE");
 	}
